plotting: Fix check_plots.py
Right now it has an issue which leads to false negatives. As it 
currently breaks the inner testing loop as soon as one challange did not 
result in valid or even in a proof at all. But to determine if a plot is 
valid i think it should try all provided challenges first and then 
decide if the plot is valid by checking if there were any proofs 
available for the given challenges, no?

So this commit removes caught_exception and its check/break if an 
exception occurs. Instead it just prints the error as debug message and 
it increases total_proofs in case of success.