Add back logging for added coins
It was removed in commit ee3dfb9eb19d238aced93daddf1f37502a368f87 for
unknown reason. A few people on keybase have learned to rely on it to
check for new coins being received by their wallet. Arguably not the
best way to do so but seems like a useful log output to have in general.