chia/cmds: separate model from view in chia farm summary
This is a first step towards making the data accessible via an API.
I'm already using a rougher version of this that is parsing the `chia farm summary`
output in a web server and rendering the data as JSON for Grafana.

Somewhat related to https://github.com/Chia-Network/chia-blockchain-gui/issues/266 as well.