"chia farm summary" remote harvester info
Info about plots is obtained from the farmer via
harvesters' farmer_info messages, instead of by
contacting a local harvester.

When multiple harvesters are in use this gives an
estimate of time-to-win based on all plots, not
just local plots. It also makes it easy to
confirm that the harvesters are communicating
with the farmer.