Add points found 24h to CLI reporting
The points_found_24h seems to include more data than that so first it
gets filtered down to the past 24 hours.  This additionally affects the
existing percent successful points 24 hours number.  Which is actually
the percent of partials successful, not points.  This could differ a bit
if the difficulty is changed during the relevant period.