Fetch file stat() before opening it in DiskProver
I don't know what is causing it, but with my plots mounted on a NAS over SMB I sometimes get a strange condition where file_path.stat() returns incorrect values (it gives the current system time for the st_mtime and st_ctime).

I can easily recreate this by a python script which opens a file in 'r' mode. After closing the file which was never written to (not even opened with 'w', '+', etc, ) an immediate call to stat() gives the current time rather then the true modified time. Waiting a second or so and then calling stat() gives the correct value.

I don't know if this is an issue with my NAS, SMB, or what but in any case this causes issues for Chia when detecting new plots. The DiskPlotter opens the file for reading, which causes the immediately subsequent call the .stat() to get the wrong value for st_mtime when adding the file to self.plots[], when the file is scanned again and the correct time is found the code on about L275 `if stat_info.st_mtime == self.plots[file_path].time_modified:` fails to match, ultimately a warning in the logs appears about a duplicate plot existing but the warning is strange as it will say that the duplicate is in the same directory as the original and neither one actually gets added to the scanned plots.

I can fix this by adding the samba directive `,actimeo=0` in my mount, but this causes the NAS to run a little slow.

I can also fix this just by moving the line calling .stat() to above the DiskPlotter so that it always gets the correct value. Since the DiskPlotter does not actually change anything in the file I don't see any negative effect of this.