Avoid activating venv in Install.ps1
PowerShell scripts, as well as batch files, operate in the callers environment which is unlike bash where you have to `source` (or `.`) scripts for that to happen.  As such, this script returned control to the user with the venv still activated which was somewhat counter to the instructions that were printed out.  I suggest this route of leaving the environment not activated on exit for the sake of consistency with `install.sh`.  I chose to simply not activate the venv to begin with since then there's no deactivation "cleanup" to manage.