Fix long load times when selecting the default wallet key from the GUI. (#8938)
When creating a new WalletNode at launch, the fingerprint will be set to None. After a lengthy initialization task, the logged-in fingerprint was set to None instead of the default private key's fingerprint. If the user then selected that same key in the key selection screen, the same lengthy initialization task would be repeated. This fix will set the logged-in fingerprint to the default key's fingerprint if the fingerprint wasn't originally specified.