rename variable to avoid hint confusion
Earlier in the function `peak_hash` is used and is detectible as
`bytes32`.  At this location it is assigned from an `Any` hinted
value thus leaving `bytes32` as a valid, although inaccurate,
hint.  Renamed to avoid this collision of uses.  When we ban `Any`
we will end up dealing with this.