Fix import for TypedDict and set request_id to bytes32
The existing import had a typo which lead mypy to treat `TypedDict`,
and thus `WsRpcMessage` as `Any`.  This allowed the hinting of
`.request_id` as `str` to slide through.  It seems to mostly be
used like `bytes32` but there is an initialization that sets it to
a `str` via `token_bytes().hex()`.  This results in the following
mypy error that will need to be considered and addressed.

```
error: Incompatible types (expression has type "str", TypedDict item "request_id" has type "bytes32")  [typeddict-item]
```