Crawler retry db on lock (#9562)
* Remove initial writes to DB, since there is nothing new to write at this point

Both of these methods are called later after each crawler iteration. The lines being deleted write all the same data back to the DB right after the initial data was loaded FROM the DB in the "create" method call, which is a waste of time, especially for mainnet where there are hundreds of thousands of records to write back

* Add additional logs around DB writes

* Add a retry around writing to the DB in case something else has a lock on the DB

Metrics processes read from this DB periodically and sometimes things time out just right and we get a lock exception. Without this retry, we'd hit the other exception handler which doesn't allow the loop to continue