test: Skip "empty" directories in `build-init-files.py`
Empty ones or such with only ignored stuff inside. It's just annoying if 
you have a branch which adds a new directory with new files and you 
switch to a branch which doesn't have those new files, running 
`pre-commit` always adds the init files in the new directory from the 
other branch because the folder from the other branch stays there and is 
empty.