wallet: Don't add derived puzzle hashes to the `WalletInterestedStore`
We already use the derived puzzle hashes when we subscribe to the node 
here 
https://github.com/Chia-Network/chia-blockchain/blob/39d733599d170b340b0ae4c13ada8896a78a0d29/chia/wallet/wallet_node.py#L1291. 
I assume this was done to make sure we subscribe to new puzzle hashes 
but at the end it leads to redundant entries in `derivation_paths` vs 
`interested_puzzle_hashes` so we can just subscribe here instead? If we 
want to keep them in both tables for some reason we may at least get rid 
of `get_puzzle_hashes_to_subscribe` and use the ones from the interest 
store only and maybe adjust the other related methods?