wallet: Avoid duplicates in `get_puzzle_hashes_to_subscribe`
We currently store the derived puzzle hashes in both, puzzle store and
interested store and here we combine them as list. There is some code to
mitigate duplicate subscriptions but its based on chunks here
https://github.com/Chia-Network/chia-blockchain/blob/39d733599d170b340b0ae4c13ada8896a78a0d29/chia/wallet/wallet_node.py#L662
so if there are less than 1000 puzzle hashes it obviously still can lead
to redundant subscription calls?