Add a marshaller for RPC APIs (#16865)
The handling of RPC requests and responses is pretty scattered and prone
to error right now. This change proposes a way to specify concrete types
for both the request and response of an RPC and have the parsing of
dictionaries occur auto-magically due to a decorator. Something similar
exists for data layer RPCs but this proposal tries to focus heavily on
the ergonomics of writing a new API and therefore handles more behavior
behind the scenes and leans on the `Streamable` library for the business
logic.

Writing a new RPC should now look like this:
```
<in some other file probably>
@streamable
@dataclass(frozen=True)
class MyNewRPCRequest(Streamable):
  class_var_1: SomeType

@streamable
@dataclass(frozen=True)
class MyNewRPCResponse(Streamable):
  class_var_2: SomeOtherType

<in *_rpc_api.py>
@marshal
async def new_rpc(self, request: MyNewRPCRequest) -> MyNewRPCResponse:
  ...
```