Delete chia/simulator/simulator_constants.py
the keychain seems unneeded, it might be wrong given the 'new consensus' comment, and it prints a python literal not a hex string or somesuch.  and it uses the `DEFAULT_ROOT_PATH` global i'm trying to get rid of.  oh, plus it's a directly runnable file that ought to be available via some `chia` subcommand instead, if it sticks around.